2nd v

+++ getchar-count.c
@@ -4,7 +4,7 @@
 
 void main()
 {
-    double nc;
+    int nc;
     for (nc = 0; getchar() != EOF; ++nc)
     ;
     printf("%.0f\n", nc);
